Move all feature selection classes to feature_selection package.
