+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/feature_selection/FeatureSelectionController.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot;
+package com.novoda.tpbot.feature_selection;
 
 public interface FeatureSelectionController<LIST, FEATURE> {
