+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/landing/LandingActivity.java
@@ -7,8 +7,8 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 
-import com.novoda.tpbot.FeatureSelectionController;
-import com.novoda.tpbot.MenuFeatureSelectionController;
+import com.novoda.tpbot.feature_selection.FeatureSelectionController;
+import com.novoda.tpbot.feature_selection.MenuFeatureSelectionController;
 import com.novoda.tpbot.R;
 import com.novoda.tpbot.bot.BotActivity;
 import com.novoda.tpbot.human.HumanActivity;
